refactor(mobile): extract voice settings and troubleshoot screens
Move the voice-settings body onto a VoiceSettingsOperations prop, lift the
troubleshoot diagnostic run into a hook behind a presentational view, and put
the network-diagnostics body behind a device-operations prop. The routes keep
the native bindings (host client acquisition, log store subscription) and mount
the screens. No visual change.
