+++ mobile/src/diagnostics/connection-diagnostics-view.tsx
@@ -1,0 +1,117 @@
+import type { ReactNode } from 'react'
+import { View, Text, Pressable } from 'react-native'
+import { useSafeAreaInsets } from 'react-native-safe-area-context'
+import { ChevronLeft, Copy, Check, Send } from 'lucide-react-native'
+import { colors, spacing } from '../theme/mobile-theme'
+import { ConnectionLog } from '../components/ConnectionLog'
+import { connectionDiagnosticsScreenStyles as styles } from './connection-diagnostics-screen-styles'
+import type { ConnectionLogEntry, ConnectionState } from '../transport/types'
+import type { ConnectionDiagnosis } from './connection-diagnostics-analysis'
+import type { DiagnosticsSubmissionState } from './connection-diagnostics-screen-data'
+
+export function ConnectionDiagnosticsView({
+  hostPicker,
+  hostName,
+  state,
+  reconnectAttempts,
+  copied,
+  copyDiagnostics,
+  diagnosis,
+  submissionState,
+  sendDiagnostics,
+  entries,
+  onBack
+}: {
+  hostPicker?: ReactNode
+  hostName: string | null
+  state: ConnectionState
+  reconnectAttempts: number
+  copied: boolean
+  copyDiagnostics: () => Promise<void>
+  diagnosis: ConnectionDiagnosis | null
+  submissionState: DiagnosticsSubmissionState | 'idle'
+  sendDiagnostics: () => Promise<void>
+  entries: readonly ConnectionLogEntry[]
+  onBack: () => void
+}) {
+  const insets = useSafeAreaInsets()
+  return (
+    <View style={[styles.container, { paddingTop: insets.top + spacing.sm }]}>
+      <View style={styles.topRow}>
+        <Pressable
+          accessibilityRole="button"
+          accessibilityLabel="Back"
+          style={styles.backButton}
+          onPress={onBack}
+        >
+          <ChevronLeft size={22} color={colors.textSecondary} />
+        </Pressable>
+        <Text style={styles.heading}>Network diagnostics</Text>
+      </View>
+
+      {hostPicker}
+      {hostName ? (
+        <>
+          <View style={styles.statusRow}>
+            <Text style={styles.statusText}>
+              {state}
+              {reconnectAttempts > 0 ? ` · attempt ${reconnectAttempts}` : ''}
+            </Text>
+            <Pressable style={styles.copyButton} onPress={() => void copyDiagnostics()}>
+              {copied ? (
+                <Check size={14} color={colors.statusGreen} />
+              ) : (
+                <Copy size={14} color={colors.textSecondary} />
+              )}
+              <Text style={styles.copyButtonText}>{copied ? 'Copied' : 'Copy report'}</Text>
+            </Pressable>
+          </View>
+          {diagnosis && (
+            <View style={styles.diagnosisCard}>
+              <Text style={styles.diagnosisHeading}>What this suggests</Text>
+              <Text style={styles.diagnosisText}>{diagnosis.likelyCause}</Text>
+              <Text style={styles.diagnosisNext}>{diagnosis.nextStep}</Text>
+              {diagnosis.reportability === 'orca-relay' && (
+                <>
+                  <Text style={styles.privacyHint}>
+                    Sends a size-limited redacted report including host name, endpoint, versions,
+                    connection state, and events—never terminal contents or credentials.
+                  </Text>
+                  <Pressable
+                    style={styles.sendButton}
+                    onPress={() => void sendDiagnostics()}
+                    disabled={submissionState === 'sending'}
+                  >
+                    {submissionState === 'sent' ? (
+                      <Check size={14} color={colors.statusGreen} />
+                    ) : (
+                      <Send size={14} color={colors.textPrimary} />
+                    )}
+                    <Text style={styles.sendButtonText}>
+                      {submissionState === 'sending'
+                        ? 'Sending…'
+                        : submissionState === 'sent'
+                          ? 'Diagnostics sent'
+                          : submissionState === 'failed'
+                            ? 'Retry sending'
+                            : 'Send diagnostics to Orca'}
+                    </Text>
+                  </Pressable>
+                </>
+              )}
+            </View>
+          )}
+          {entries.length > 0 ? (
+            <ConnectionLog entries={[...entries]} title={hostName} fillAvailableHeight />
+          ) : (
+            <Text style={styles.emptyText}>
+              No connection events yet. Events appear as the app dials this host.
+            </Text>
+          )}
+        </>
+      ) : (
+        <Text style={styles.emptyText}>No paired hosts.</Text>
+      )}
+    </View>
+  )
+}
