+++ mobile/src/diagnostics/diagnostics-device-operations.ts
@@ -1,0 +1,16 @@
+import type {
+  ConnectionDiagnosticsSubmission,
+  ConnectionDiagnosticsSubmissionResult
+} from './connection-diagnostics-submission'
+
+/** A fresh report plus the incident it describes, so a stale send can be dropped. */
+export type ConnectionDiagnosticsReport = ConnectionDiagnosticsSubmission & {
+  incidentId: string | null
+}
+
+export interface DiagnosticsDeviceOperations {
+  report(): Promise<ConnectionDiagnosticsReport>
+  submit(
+    submission: ConnectionDiagnosticsSubmission
+  ): Promise<ConnectionDiagnosticsSubmissionResult>
+}
