+++ mobile/src/settings/voice-settings-operations.ts
@@ -1,0 +1,12 @@
+import type { MobileSpeechSetup } from '../dictation/mobile-dictation-setup'
+
+export interface VoiceSettingsOperations {
+  load(): Promise<MobileSpeechSetup>
+  configure(params: {
+    enabled?: boolean
+    modelId?: string
+    dictationMode?: 'toggle' | 'hold'
+  }): Promise<MobileSpeechSetup>
+  download(modelId: string): Promise<void>
+  delete(modelId: string): Promise<MobileSpeechSetup>
+}
